Use one parsingHook for SQL and DistSQL parser to sent tracing event
