+++ shardingsphere-infra/shardingsphere-infra-parser/src/main/java/org/apache/shardingsphere/infra/parser/ShardingSphereSQLParserEngine.java
@@ -20,6 +20,8 @@ package org.apache.shardingsphere.infra.parser;
 import org.apache.shardingsphere.distsql.parser.DistSQLStatementParserEngine;
 import org.apache.shardingsphere.infra.parser.sql.SQLStatementParserEngine;
 import org.apache.shardingsphere.infra.parser.sql.SQLStatementParserEngineFactory;
+import org.apache.shardingsphere.sql.parser.exception.SQLParsingException;
+import org.apache.shardingsphere.sql.parser.hook.ParsingHookRegistry;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 
 /**
@@ -31,11 +33,19 @@ public final class ShardingSphereSQLParserEngine {
     
     private final DistSQLStatementParserEngine distSQLStatementParserEngine;
     
+    private final ParsingHookRegistry parsingHookRegistry;
+    
     public ShardingSphereSQLParserEngine(final String databaseTypeName) {
         sqlStatementParserEngine = SQLStatementParserEngineFactory.getSQLStatementParserEngine(databaseTypeName);
         distSQLStatementParserEngine = new DistSQLStatementParserEngine();
+        parsingHookRegistry = ParsingHookRegistry.getInstance();
     }
     
+    /*
+     * To make sure SkyWalking will be available at the next release of ShardingSphere, a new plugin should be provided to SkyWalking project if this API changed.
+     *
+     * @see <a href="https://github.com/apache/skywalking/blob/master/docs/en/guides/Java-Plugin-Development-Guide.md#user-content-plugin-development-guide">Plugin Development Guide</a>
+     */
     /**
      * Parse to SQL statement.
      *
@@ -43,15 +53,31 @@ public final class ShardingSphereSQLParserEngine {
      * @param useCache whether use cache
      * @return SQL statement
      */
+    @SuppressWarnings("OverlyBroadCatchBlock")
     public SQLStatement parse(final String sql, final boolean useCache) {
-        SQLStatement result;
+        parsingHookRegistry.start(sql);
         try {
-            result = sqlStatementParserEngine.parse(sql, useCache);
+            SQLStatement result = parse0(sql, useCache);
+            parsingHookRegistry.finishSuccess(result);
+            return result;
             // CHECKSTYLE:OFF
+            // TODO check whether throw SQLParsingException only
         } catch (final Exception ex) {
             // CHECKSTYLE:ON
-            result = distSQLStatementParserEngine.parse(sql);
+            parsingHookRegistry.finishFailure(ex);
+            throw ex;
         }
-        return result;
+    }
+    
+    private SQLStatement parse0(final String sql, final boolean useCache) {
+        try {
+            return sqlStatementParserEngine.parse(sql, useCache);
+        } catch (final SQLParsingException originalEx) {
+            try {
+                return distSQLStatementParserEngine.parse(sql);
+            } catch (final SQLParsingException ignored) {
+                throw originalEx;
+            }
+        }
     }
 }
